Return hoverintent object from options
This tripped me up big time at work yesterday: I was using hoverintent
happily, and was using an array to save references to hoverintent
objects so I could remove them later.

I later added an `.options` call to adjust the sensitivity, and
discovered when I came in this morning that things had broken badly:
event listeners were piling up because they were no longer being removed
properly. I discovered that it was because as soon as I added
`.options`, the reference I was saving was undefined.

This commit also adds the invaluable `.remove()` to the README.